+++ R/read_wave.R
@@ -63,7 +63,7 @@ read_wave <- function (X, index, from = X$start[index], to = X$end[index], chann
   if (any(is.na(c(X$end, X$start)))) stop("NAs found in start and/or end")  
   
   #if end or start are not numeric stop
-  if (all(class(X$end) != "numeric" & class(X$start) != "numeric")) stop("'start' and 'end' must be numeric")
+  if (any(!is(X$end, "numeric"), !is(X$start, "numeric"))) stop("'start' and 'end' must be numeric")
   
   #if any start higher than end stop
   if (any(X$end - X$start < 0)) stop(paste("The start is higher than the end in", length(which(X$end - X$start<0)), "case(s)"))